+++ tests_old/promising_context/sync/test_sync_context_decorator.py
@@ -2,96 +2,6 @@ import pytest
 
 import promising
 
-# ── Sync Function Decorator ──────────────────────────────────────
-
-
-async def test_context_decorator_activates_context() -> None:
-    """
-    @promising.context on a sync function: the context is
-    active inside the function body.
-    """
-    captured_ctx = None
-
-    @promising.context
-    def work() -> str:
-        nonlocal captured_ctx
-        captured_ctx = promising.get_active_context()
-        return "sync-done"
-
-    assert work() == "sync-done"
-    assert captured_ctx is not None
-    assert isinstance(captured_ctx, promising.PromisingContext)
-
-
-async def test_context_decorator_deactivates_after() -> None:
-    """
-    After the decorated sync function returns, the context is
-    no longer active.
-    """
-
-    @promising.context
-    def work() -> str:
-        return "done"
-
-    assert promising.get_active_context(raise_if_none=False) is None
-    work()
-    assert promising.get_active_context(raise_if_none=False) is None
-
-
-async def test_context_decorator_forwards_args() -> None:
-    """
-    Positional and keyword arguments are forwarded to the
-    decorated sync function.
-    """
-
-    @promising.context
-    def add(a: int, b: int, *, multiplier: int = 1) -> int:
-        return (a + b) * multiplier
-
-    assert add(3, 4) == 7
-    assert add(3, 4, multiplier=2) == 14
-
-
-async def test_context_decorator_exception_propagates() -> None:
-    """
-    An exception raised inside the decorated sync function
-    propagates to the caller.
-    """
-
-    @promising.context
-    def failing() -> None:
-        raise ValueError("sync func error")
-
-    with pytest.raises(ValueError, match="sync func error"):
-        failing()
-
-
-async def test_context_decorator_deactivates_on_exception() -> None:
-    """
-    The context is deactivated even if the decorated sync function
-    raises.
-    """
-
-    @promising.context
-    def failing() -> None:
-        raise RuntimeError("boom")
-
-    with pytest.raises(RuntimeError, match="boom"):
-        failing()
-
-    assert promising.get_active_context(raise_if_none=False) is None
-
-
-async def test_context_decorator_with_parens() -> None:
-    @promising.context()
-    def work() -> str:
-        return "parens-sync"
-
-    assert work() == "parens-sync"
-
-
-# ── Sync Instance Methods ────────────────────────────────────────
-
 
 async def test_sync_instance_method_activates_context() -> None:
     """
